Review of data_analysis.py
